Deprecate `Yoast_ACF_Analysis_Facade::get_filter_name()`
Having filter names created via a function call makes it impossible for documentation scrapers to document the available filters.

This commit deprecated the method which was used for this and changes all existing filter to hard-coded filter names.

The filter names, however, are still the same, so this does not constitute a BC-break.