Do not expect Gutenberg to be available
We cannot access the core/block-editor store if someone is using the Classic Editor. Therefore, I have added an additional check to see whether the selector is available.

If not, we fallback to the method that we used earlier.

Although this is not the nicest solution, it works. And I do not have enough context to actually create something different here.