Travis: run the tests against PHP 8/nightly
Now the tests are fully compatible with PHPUnit 9, the test suite can be run on PHP 8.

To that end, we run a `composer update` for just and only the test dependencies on PHP 8 to make sure that the test dependencies available are compatible with PHP 8.
And as the `composer.json` contains a hard-coded `platform.php` setting, we use `ignore-platform-reqs`  to get round that setting.

Includes:
* Adding `travis_retry` to `composer install`-like commands to get round builds stalling on the connection with Packagist, especially on older PHP versions.
* Removing `--prefer-dist` from the `composer install` command as that is the default, so doesn't need to be explicitly passed.