Tests: selectively switch over to the Yoast\PHPUnitPolyfills\TestCases\TestCase
For those test classes in the ACF testsuite which are not based on BrainMonkey, we don't need the overhead of loading the BrainMonkey framework, but do want the PHPUnit cross-version compatibility.

So, the `use` statements referring to the PHPUnit native `PHPUnit\Framework\TestCase` have been replaced with `use` statements for the PHPUnit cross-version compatible `Yoast\PHPUnitPolyfills\TestCases\TestCase`.